pyproject.toml
[project]
name = "rlane-fidelity"
version = "1.0.1"
description = "Process Fidelity transaction files"
authors = [
    {name = "Russel Lane", email = "russel@rlane.com"},
]
license = {text = "MIT"}
readme = "README.md"
keywords = ["fidelity", "python", "trading", "stocks", "reports"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial",
]
requires-python = ">=3.10"
dependencies = [
    "rich>=13.9.4",
    "rlane-libcli>=1.0.8",
]

[project.urls]
Homepage = "https://github.com/russellane/fidelity"

[project.scripts]
fidelity = 'fidelity.cli:main'

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.10.0",
    "flake8-bugbear>=24.10.31",
    "flake8-pydocstyle>=0.2.4",
    "flake8-pylint>=0.2.1",
    "flake8-pytest-style>=2.0.0",
    "flake8-simplify>=0.21.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.3",
]

[tool.black]
line-length = "97"

[tool.isort]
line_length = "97"
profile = "black"

[tool.pylint.format]
max-line-length = "97"

[tool.pylint.messages_control]
enable = [
    "useless-suppression",
]
fail-on = [
    # useless-suppression
    "I0021",
]

[tool.pydocstyle]
convention = "google"
add-ignore = [
    # "D105",     # pep257 numpy google   # Missing docstring in magic method
    "D105",
    # "D202",     # pep257 numpy google   # No blank lines allowed after function docstring
    "D202",
]

[tool.mypy]
strict = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# vim: set ts=4 sw=4 et:
